+++ Todo-list/js/modules/main.js
@@ -1,7 +1,12 @@
 import {getLocalData, setLocalData, removeLocalData, valideLocalData} from './userData.js'
-import {cssWrapper, cssPanel, cssPanelButton, cssPanelInput, cssRoot} from './style.js'
-import {setHover, setFocus} from './events.js';
-import {renderCard} from './card.js';
+import {cssCard, cssCardStatusWraper, cssStatusSpan, cssCardStatus, cssCardDate, cssCardText, cssCardCloseButton, cssImage} from './style.js'
+import {setHover, setFocus,setCardStyle} from './events.js';
+import {
+    root, wrapper,
+    panelActions, buttonDeleteAllCards,buttonDeleteLastCard, inputTask, buttonAddCard,
+    panelCardInfo, counterAllCards, counterCompletedCards, buttonShowAll, buttonShowCompleted, inputSearch,
+    cardsAera
+}from './render.js'
 
 export let main = function (){
 
@@ -16,7 +21,6 @@ export let main = function (){
 
     let completed = main.filter(item => item.status);
     let searched = [];
-
 
 const deleteAllCards = function(){
     let result = confirm("Delete All Cards?");
@@ -57,8 +61,6 @@ const deleteLastCard = function(){
     }
 };
 
-
-
 const setCounterAllCards = function(){
     counterAllCards.innerHTML = `All: <span>${main.length}</span>`;
 }
@@ -76,6 +78,78 @@ const searchForCompleted = function(){
     searched = completed.filter(item => item.task.includes(this.value.trim()));
     createCardsAera(searched);
 };
+
+const renderCard = function (item) {
+    let card = document.createElement("div");
+        card.classList.add("card");
+        card.style.cssText = cssCard;
+        card.setAttribute("data-id", item.id);
+
+    let cardStatusWraper = document.createElement("label");
+        cardStatusWraper.style.cssText = cssCardStatusWraper;
+
+    let cardStatusSpan = document.createElement("span");
+        cardStatusSpan.style.cssText = cssStatusSpan;
+
+    let cardStatus = document.createElement("input");
+        cardStatus.setAttribute("type", "checkbox");
+        cardStatus.checked = item.status;
+        cardStatus.classList.add("card--status");
+        cardStatus.style.cssText = cssCardStatus;
+        if (item.status){
+            card.style.backgroundColor= "gray";
+            cardStatusSpan.style.backgroundImage = cssImage;
+
+            let key =  completed.some( elem => elem.id === item.id);
+            if (key == false){
+                completed.push(item);
+            }
+        }
+        setFocus(cardStatus, cardStatusSpan, "0 0 0 0.15em DodgerBlue")
+        setFocus(cardStatus, card, "0 0 0 0.04em black")
+
+    let cardDate = document.createElement("div");
+        cardDate.classList.add("card--date");
+        cardDate.textContent = item.date
+        cardDate.style.cssText = cssCardDate
+
+    let cardText = document.createElement("div");
+        cardText.classList.add("card--text");
+        cardText.textContent = item.task
+        cardText.style.cssText = cssCardText
+        if (item.status){
+            cardText.style.textDecoration = "line-through";
+        }
+
+    let cardCloseButton = document.createElement("div");
+        cardCloseButton.classList.add("card--date");
+        cardCloseButton.textContent = "X"
+        cardCloseButton.style.cssText = cssCardCloseButton;
+        cardCloseButton.addEventListener("click", deleteCard);
+        setHover("Red", "0.4", cardCloseButton);
+        
+
+        cardStatus.addEventListener("click",setCardStyle);
+        cardStatus.addEventListener("click", function(){
+             item.status = cardStatus.checked;
+             if(item.status){
+                completed.push(item);
+             }else{
+                let parent = this.closest(".card")
+                let newCompleted = completed.filter(item => item.id != parent.getAttribute('data-id'));
+                completed.length = 0;
+                completed = [...newCompleted].sort((a,b) => b.id - a.id);
+             }
+     
+            setCounterCompletedCards();
+            setLocalData(main);
+        });
+        setFocus(cardStatus)
+
+        cardsAera.prepend(card);
+        card.append(cardStatusWraper, cardText, cardCloseButton, cardDate); 
+        cardStatusWraper.append(cardStatusSpan, cardStatus);
+    }
 
 const addCard = function(){
     
@@ -97,7 +171,7 @@ const addCard = function(){
         obj.date = `${date.getDate()} - ${date.getMonth()+1} - ${date.getFullYear()} `;
     
         main.push(obj);
-        renderCard(obj, main, completed, cardsAera, deleteCard, setCounterCompletedCards,setLocalData);
+        renderCard(obj);
         inputTask.value = "";
         setLocalData(main);
         setCounterAllCards();
@@ -126,74 +200,24 @@ const deleteCard = function(){
 
 const createCardsAera = function (arr){ 
     arr.forEach((item) => {
-        renderCard(item, main, completed, cardsAera, deleteCard, setCounterCompletedCards,setLocalData);
+        renderCard(item);
     })
     setCounterAllCards();
     setCounterCompletedCards(); 
 }
 
-function renderApp(){
-    
-}
-let root = document.getElementById("root")
-    root.style.cssText = cssRoot;
- 
-let wrapper = document.createElement("div");
-    wrapper.classList.add("wrapper");
-    wrapper.style.cssText =  cssWrapper;
- 
-let panelActions = document.createElement("div");
-    panelActions.classList.add("panel-card-actions");
-    panelActions.style.cssText = cssPanel;
-
-let buttonDeleteAllCards = document.createElement("button");
-    buttonDeleteAllCards.textContent = "Delete All";
-    buttonDeleteAllCards.style.cssText = cssPanelButton;
     buttonDeleteAllCards.addEventListener("click", deleteAllCards);
-    setHover("Red","0.4", buttonDeleteAllCards);
-    setFocus(buttonDeleteAllCards);
-
-let buttonDeleteLastCard = document.createElement("button");
-    buttonDeleteLastCard.textContent = "Delete last";
-    buttonDeleteLastCard.style.cssText = cssPanelButton;
     buttonDeleteLastCard.addEventListener("click", deleteLastCard);
-    setHover("Red","0.4", buttonDeleteLastCard);
-    setFocus(buttonDeleteLastCard);
-
-
-let inputTask = document.createElement("input");
-    inputTask.setAttribute("type", "text");
-    inputTask.setAttribute("placeholder", "Enter todo...");
-    inputTask.style.cssText = cssPanelInput;
     inputTask.addEventListener("keyup", function(event) {
         if (event.key === "Enter") {
             addCard();
         }
     })
-    setFocus(inputTask);
-
-let buttonAddCard = document.createElement("button");
-    buttonAddCard.textContent = "Add";
-    buttonAddCard.style.cssText = cssPanelButton;
     buttonAddCard.addEventListener("click", addCard);
-    setHover("yellowgreen","0.4", buttonAddCard);
-    setFocus(buttonAddCard);
 
-
-let panelCardInfo = document.createElement("div");
-    panelCardInfo.classList.add("panel-card-actions");
-    panelCardInfo.style.cssText=cssPanel;
-
-
-let counterAllCards= document.createElement("div");
     setCounterAllCards();
-
-let counterCompletedCards = document.createElement("div");
     setCounterCompletedCards();
 
-let buttonShowAll = document.createElement("button");
-    buttonShowAll.textContent = "Show All";
-    buttonShowAll.style.cssText = cssPanelButton;
     buttonShowAll.addEventListener("click",function(){
         cardsAera.innerHTML = "";
         createCardsAera(main);
@@ -201,12 +225,7 @@ let buttonShowAll = document.createElement("button");
         inputSearch.removeEventListener("keyup",searchForCompleted);
         inputSearch.addEventListener("keyup", searchForMain);
     })
-    setHover("yellowgreen","0.4", buttonShowAll);
-    setFocus(buttonShowAll);
 
-let buttonShowCompleted = document.createElement("button");
-    buttonShowCompleted.textContent = "Show Completed";
-    buttonShowCompleted.style.cssText = cssPanelButton;
     buttonShowCompleted.addEventListener("click", function(){
         cardsAera.innerHTML = "";
         createCardsAera(completed);
@@ -214,17 +233,6 @@ let buttonShowCompleted = document.createElement("button");
         inputSearch.removeEventListener("keyup",searchForMain);
         inputSearch.addEventListener("keyup", searchForCompleted);       
     }); 
-    setHover("yellowgreen","0.4", buttonShowCompleted);
-    setFocus(buttonShowCompleted);
-
-let inputSearch = document.createElement("input");
-    inputSearch.setAttribute("type", "text");
-    inputSearch.setAttribute("placeholder", "Search..."); 
-    inputSearch.style.cssText = cssPanelInput;
-    setFocus(inputSearch);
-
-let cardsAera = document.createElement("div");
-    cardsAera.classList.add("cards-aera");
 
 
  createCardsAera(main);
